requirements.txt
django
django-crispy-forms
djangorestframework
django-cors-headers
